+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning].
 
 ### Features
 
+- Maintain capitalization of all words in a mapping phrase. ([#26])
 - Add option to specify a mapping from the CLI. ([#31])
 
 ### Bug Fixes
@@ -51,6 +52,7 @@ Versioning].
 [#22]: https://github.com/ericcornelissen/wordrow/pull/22
 [#21]: https://github.com/ericcornelissen/wordrow/pull/21
 [#23]: https://github.com/ericcornelissen/wordrow/pull/23
+[#26]: https://github.com/ericcornelissen/wordrow/pull/26
 [#29]: https://github.com/ericcornelissen/wordrow/pull/29
 [#30]: https://github.com/ericcornelissen/wordrow/pull/30
 [#31]: https://github.com/ericcornelissen/wordrow/pull/31